# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Nhựa API",
        "description": "API quản lý sản phẩm và biến thể sản phẩm",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/variant-options": {
            "get": {
                "tags": [
                    "Variant Options"
                ],
                "summary": "Danh sách giá trị biến thể",
                "operationId": "7640f9b76f2fc5fdbb1204651aa2f681",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "variant_group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Variant Options"
                ],
                "summary": "Tạo giá trị biến thể",
                "operationId": "6555c3321249ae02fec94cbc99f200a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVariantOptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đã tạo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VariantOptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/variant-options/{id}": {
            "get": {
                "tags": [
                    "Variant Options"
                ],
                "summary": "Chi tiết giá trị biến thể",
                "operationId": "5d7cac23e897da3d280f51470abdca00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VariantOptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Variant Options"
                ],
                "summary": "Cập nhật giá trị biến thể",
                "operationId": "81e3be6229077cba87d6c39a0aeae13e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVariantOptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đã cập nhật",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/VariantOptionResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Variant Options"
                ],
                "summary": "Xóa giá trị biến thể",
                "operationId": "beded209efba32474ce376c9133a8458",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã xóa"
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreVariantOptionRequest": {
                "required": [
                    "variant_group_id",
                    "option_code",
                    "option_name"
                ],
                "properties": {
                    "variant_group_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "option_code": {
                        "type": "string",
                        "example": "red",
                        "maxLength": 100
                    },
                    "option_name": {
                        "type": "string",
                        "example": "Màu đỏ",
                        "maxLength": 255
                    },
                    "sort_order": {
                        "type": "integer",
                        "nullable": true,
                        "minimum": 0
                    },
                    "is_active": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateVariantOptionRequest": {
                "properties": {
                    "variant_group_id": {
                        "type": "integer"
                    },
                    "option_code": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "option_name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "sort_order": {
                        "type": "integer",
                        "nullable": true,
                        "minimum": 0
                    },
                    "is_active": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "VariantOptionResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "variant_group_id": {
                        "type": "integer"
                    },
                    "option_code": {
                        "type": "string"
                    },
                    "option_name": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "integer"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer token",
                "scheme": "bearer"
            }
        }
    },
...
}